+++ source/job/job_place.c
@@ -47,29 +47,26 @@ static job_place_s jp;
 
 /* Z profiles */
 static const motion_profile_s g_place_z_down_profile = PLACE_Z_DOWN_PROFILE;
-static const motion_profile_s g_place_z_up_profile   = PLACE_Z_UP_PROFILE;
+static const motion_profile_s g_place_z_up_profile = PLACE_Z_UP_PROFILE;
 
 /* -------------------------------------------------------------------------- */
 /* Helpers                                                                    */
 /* -------------------------------------------------------------------------- */
 
-static bool wait_elapsed(uint32_t start_tick, uint32_t wait_ticks)
-{
-    return (motion_get_isr_tick_count() - start_tick) >= wait_ticks;
+static bool wait_elapsed(uint32_t start_tick, uint32_t wait_ticks){
+    return (uint32_t)(motion_get_isr_tick_count() - start_tick) >= wait_ticks;
 }
 
-static void place_enter_wait(place_state_e next_state)
-{
+static void place_enter_wait(place_state_e next_state){
     jp.state = next_state;
     jp.wait_start_tick = motion_get_isr_tick_count();
 }
 
-static bool place_fail(err_e e, err_e *out_err)
-{
+static bool place_fail(err_e e, err_e* out_err){
     jp.state = PLACE_STATE_ERROR;
     jp.last_err = e;
 
-    if (out_err != 0) {
+    if(out_err != 0){
         *out_err = e;
     }
     return true;
@@ -83,14 +80,12 @@ static bool place_fail(err_e e, err_e *out_err)
  * Therefore:
  *   current_z > safe_z => tool is too low, so first move only Z to safe.
  */
-static bool place_needs_raise_to_safe(void)
-{
+static bool place_needs_raise_to_safe(void){
     return g_status.pos_internal.z_mm_scaled > PLACE_Z_SAFE_POS_MM_SCALED;
 }
 
-static err_e place_start_raise_to_safe_z(void)
-{
-    bot_action_s cmd = {0};
+static err_e place_start_raise_to_safe_z(void){
+    bot_action_s cmd = { 0 };
 
     cmd.type = ACT_PLACE;
     cmd.target_pos = g_status.pos_internal;
@@ -99,17 +94,16 @@ static err_e place_start_raise_to_safe_z(void)
     return motion_start(&cmd, limit_none, &g_place_z_up_profile);
 }
 
-static err_e place_start_xy_phi_move(const robot_pos_s *target_pos)
-{
-    bot_action_s cmd = {0};
+static err_e place_start_xy_phi_move(const robot_pos_s* target_pos){
+    bot_action_s cmd = { 0 };
 
     cmd.type = ACT_PLACE;
     cmd.target_pos = g_status.pos_internal;
 
     /* Move only in XY+phi, keep Z at safe height */
-    cmd.target_pos.x_mm_scaled    = target_pos->x_mm_scaled;
-    cmd.target_pos.y_mm_scaled    = target_pos->y_mm_scaled;
-    cmd.target_pos.z_mm_scaled    = PLACE_Z_SAFE_POS_MM_SCALED;
+    cmd.target_pos.x_mm_scaled = target_pos->x_mm_scaled;
+    cmd.target_pos.y_mm_scaled = target_pos->y_mm_scaled;
+    cmd.target_pos.z_mm_scaled = PLACE_Z_SAFE_POS_MM_SCALED;
     cmd.target_pos.phi_deg_scaled = target_pos->phi_deg_scaled;
 
     jp.xy_phi_safe_target = cmd.target_pos;
@@ -118,9 +112,8 @@ static err_e place_start_xy_phi_move(const robot_pos_s *target_pos)
     return motion_start(&cmd, limit_none, 0);
 }
 
-static err_e place_start_z_down(void)
-{
-    bot_action_s cmd = {0};
+static err_e place_start_z_down(void){
+    bot_action_s cmd = { 0 };
 
     cmd.type = ACT_PLACE;
     cmd.target_pos = g_status.pos_internal;
@@ -129,9 +122,8 @@ static err_e place_start_z_down(void)
     return motion_start(&cmd, limit_none, &g_place_z_down_profile);
 }
 
-static err_e place_start_z_up(void)
-{
-    bot_action_s cmd = {0};
+static err_e place_start_z_up(void){
+    bot_action_s cmd = { 0 };
 
     cmd.type = ACT_PLACE;
     cmd.target_pos = g_status.pos_internal;
@@ -144,22 +136,21 @@ static err_e place_start_z_up(void)
 /* Public API                                                                 */
 /* -------------------------------------------------------------------------- */
 
-err_e job_place_start(const bot_action_s *a)
-{
+err_e job_place_start(const bot_action_s* a){
     err_e e;
 
-    if (a == 0) {
+    if(a == 0){
         return ERR_INTERNAL;
     }
 
-    jp = (job_place_s){0};
+    jp = (job_place_s){ 0 };
     jp.state = PLACE_STATE_IDLE;
     jp.last_err = ERR_NONE;
     jp.place_target = a->target_pos;
 
-    if (place_needs_raise_to_safe()) {
+    if(place_needs_raise_to_safe()){
         e = place_start_raise_to_safe_z();
-        if (e != ERR_NONE) {
+        if(e != ERR_NONE){
             jp.state = PLACE_STATE_ERROR;
             jp.last_err = e;
             return e;
@@ -170,7 +161,7 @@ err_e job_place_start(const bot_action_s *a)
     }
 
     e = place_start_xy_phi_move(&jp.place_target);
-    if (e != ERR_NONE) {
+    if(e != ERR_NONE){
         jp.state = PLACE_STATE_ERROR;
         jp.last_err = e;
         return e;
@@ -180,24 +171,23 @@ err_e job_place_start(const bot_action_s *a)
     return ERR_NONE;
 }
 
-bool job_place_step(err_e *out_err)
-{
+bool job_place_step(err_e* out_err){
     err_e e = ERR_NONE;
     bool done = false;
 
-    switch (jp.state) {
+    switch(jp.state){
     case PLACE_STATE_RAISE_TO_SAFE_Z:
-        if (!motion_is_done()) {
+        if(!motion_is_done()){
             return false;
         }
 
         e = motion_last_err();
-        if (e != ERR_NONE) {
+        if(e != ERR_NONE){
             return place_fail(e, out_err);
         }
 
         e = place_start_xy_phi_move(&jp.place_target);
-        if (e != ERR_NONE) {
+        if(e != ERR_NONE){
             return place_fail(e, out_err);
         }
 
@@ -206,11 +196,11 @@ bool job_place_step(err_e *out_err)
 
     case PLACE_STATE_XY_PHI_MOVE:
         done = job_motion_finish_step(&jp.xy_phi_finish, &e);
-        if (!done) {
+        if(!done){
             return false;
         }
 
-        if (e != ERR_NONE) {
+        if(e != ERR_NONE){
             return place_fail(e, out_err);
         }
 
@@ -218,12 +208,12 @@ bool job_place_step(err_e *out_err)
         return false;
 
     case PLACE_STATE_XY_PHI_SETTLE:
-        if (!wait_elapsed(jp.wait_start_tick, PLACE_WAIT_TICKS_XY_PHI_SETTLE)) {
+        if(!wait_elapsed(jp.wait_start_tick, PLACE_WAIT_TICKS_XY_PHI_SETTLE)){
             return false;
         }
 
         e = place_start_z_down();
-        if (e != ERR_NONE) {
+        if(e != ERR_NONE){
             return place_fail(e, out_err);
         }
 
@@ -231,12 +221,12 @@ bool job_place_step(err_e *out_err)
         return false;
 
     case PLACE_STATE_Z_DOWN:
-        if (!motion_is_done()) {
+        if(!motion_is_done()){
             return false;
         }
 
         e = motion_last_err();
-        if (e != ERR_NONE) {
+        if(e != ERR_NONE){
             return place_fail(e, out_err);
         }
 
@@ -244,7 +234,7 @@ bool job_place_step(err_e *out_err)
         return false;
 
     case PLACE_STATE_Z_DOWN_SETTLE:
-        if (!wait_elapsed(jp.wait_start_tick, PLACE_WAIT_TICKS_Z_SETTLE)) {
+        if(!wait_elapsed(jp.wait_start_tick, PLACE_WAIT_TICKS_Z_SETTLE)){
             return false;
         }
 
@@ -253,12 +243,12 @@ bool job_place_step(err_e *out_err)
         return false;
 
     case PLACE_STATE_MAGNET_RELEASE_WAIT:
-        if (!wait_elapsed(jp.wait_start_tick, PLACE_WAIT_TICKS_MAGNET_RELEASE)) {
+        if(!wait_elapsed(jp.wait_start_tick, PLACE_WAIT_TICKS_MAGNET_RELEASE)){
             return false;
         }
 
         e = place_start_z_up();
-        if (e != ERR_NONE) {
+        if(e != ERR_NONE){
             return place_fail(e, out_err);
         }
 
@@ -266,12 +256,12 @@ bool job_place_step(err_e *out_err)
         return false;
 
     case PLACE_STATE_Z_UP:
-        if (!motion_is_done()) {
+        if(!motion_is_done()){
             return false;
         }
 
         e = motion_last_err();
-        if (e != ERR_NONE) {
+        if(e != ERR_NONE){
             return place_fail(e, out_err);
         }
 
@@ -279,33 +269,33 @@ bool job_place_step(err_e *out_err)
         return false;
 
     case PLACE_STATE_Z_UP_SETTLE:
-        if (!wait_elapsed(jp.wait_start_tick, PLACE_WAIT_TICKS_Z_SETTLE)) {
+        if(!wait_elapsed(jp.wait_start_tick, PLACE_WAIT_TICKS_Z_SETTLE)){
             return false;
         }
 
         jp.state = PLACE_STATE_DONE;
         jp.last_err = ERR_NONE;
 
-        if (out_err != 0) {
+        if(out_err != 0){
             *out_err = ERR_NONE;
         }
         return true;
 
     case PLACE_STATE_DONE:
-        if (out_err != 0) {
+        if(out_err != 0){
             *out_err = ERR_NONE;
         }
         return true;
 
     case PLACE_STATE_ERROR:
-        if (out_err != 0) {
+        if(out_err != 0){
             *out_err = jp.last_err;
         }
         return true;
 
     case PLACE_STATE_IDLE:
     default:
-        if (out_err != 0) {
+        if(out_err != 0){
             *out_err = ERR_INTERNAL;
         }
         return true;
